refactor(picking): refactor and add picking for ImageArrayMapper
Refactor picking functions from vtkImageMapper to
AbstractImageMapper\helper and add picking functionality
to ImageArrayMapper.
